Move constants out of views directory
