+++ website/draft_site/drafts/views/__init__.py
@@ -7,5 +7,3 @@ from .index import index
 from .pick_card import pick_card
 from .show_draft import show_draft
 from .show_seat import show_seat
-# TODO: might want to move this since it's not a view
-from .constants import CUBES, CUBES_BY_ID
