+++ website/draft_site/drafts/views/all_picks.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404
 
-from .constants import CUBES_BY_ID
+from ..constants import CUBES_BY_ID
 from .. import models
